Update server and lambda js files to only onvoke via AWS SDK
